requirements.txt
numpy>=1.24
pandas>=2.0
Pillow>=9.5
PyYAML>=6.0
dvc>=3.0
torch>=2.2
diffusers>=0.27
timm>=1.0
geffnet>=1.0
mlflow>=2.0
unidecode>=1.3
